SafeApp/flask/requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
Pillow==8.1.0
pycparser==2.20
pycryptodome==3.9.9
six==1.15.0
SQLAlchemy==1.3.22
uWSGI==2.0.19.1
Werkzeug==1.0.1
WTForms==2.3.3
